refactor: move method definitions exposed in the scripting env under utils
