+++ packages/hoppscotch-js-sandbox/src/pre-request/web-worker/worker.ts
@@ -1,19 +1,14 @@
-import { parseTemplateStringE } from "@hoppscotch/data"
-import * as E from "fp-ts/Either"
-import * as O from "fp-ts/Option"
 import * as TE from "fp-ts/TaskEither"
-import { pipe } from "fp-ts/lib/function"
 import { cloneDeep } from "lodash-es"
+
 import { TestResult } from "~/types"
-import { getEnv, setEnv } from "../../utils"
+import { getPreRequestScriptMethods } from "../../utils"
 
 const executeScriptInContext = (
   preRequestScript: string,
   envs: TestResult["envs"]
 ) => {
   try {
-    let currentEnvs = cloneDeep(envs)
-
     // Create a function from the script using the Function constructor
     const scriptFunction = new Function(
       "pw",
@@ -22,87 +17,12 @@ const executeScriptInContext = (
       preRequestScript
     )
 
-    const envGetHandle = (key: any) => {
-      if (typeof key !== "string") {
-        return TE.left("Expected key to be a string")
-      }
-
-      const result = pipe(
-        getEnv(key, currentEnvs),
-        O.match(
-          () => undefined,
-          ({ value }) => String(value)
-        )
-      )
-
-      return result
-    }
-
-    const envGetResolveHandle = (key: any) => {
-      if (typeof key !== "string") {
-        return TE.left("Expected key to be a string")
-      }
-
-      const result = pipe(
-        getEnv(key, currentEnvs),
-        E.fromOption(() => "INVALID_KEY" as const),
-
-        E.map(({ value }) =>
-          pipe(
-            parseTemplateStringE(value, [...envs.selected, ...envs.global]),
-            // If the recursive resolution failed, return the unresolved value
-            E.getOrElse(() => value)
-          )
-        ),
-        E.map((x) => String(x)),
-        E.getOrElseW(() => undefined)
-      )
-
-      return result
-    }
-
-    const envSetHandle = (key: any, value: any) => {
-      if (typeof key !== "string") {
-        return TE.left("Expected key to be a string")
-      }
-
-      if (typeof value !== "string") {
-        return TE.left("Expected value to be a string")
-      }
-      currentEnvs = setEnv(key, value, currentEnvs)
-
-      return undefined
-    }
-
-    const envResolveHandle = (value: any) => {
-      if (typeof value !== "string") {
-        return TE.left("Expected value to be a string")
-      }
-
-      const result = pipe(
-        parseTemplateStringE(value, [
-          ...currentEnvs.selected,
-          ...currentEnvs.global,
-        ]),
-        E.getOrElse(() => value)
-      )
-
-      return String(result)
-    }
-
-    const pw = {
-      env: {
-        get: envGetHandle,
-        getResolve: envGetResolveHandle,
-        set: envSetHandle,
-        resolve: envResolveHandle,
-      },
-    }
+    const { pw, updatedEnvs } = getPreRequestScriptMethods(envs)
 
     // Expose pw and other dependencies to the script
-    scriptFunction(pw, cloneDeep, currentEnvs)
+    scriptFunction(pw, cloneDeep, updatedEnvs)
 
-    return TE.right(currentEnvs)
+    return TE.right(updatedEnvs)
   } catch (error) {
     return TE.left("Script execution failed: " + (error as Error).message)
   }
