+++ packages/hoppscotch-js-sandbox/src/test-runner/web-worker/worker.ts
@@ -2,20 +2,12 @@
  * Web worker based implementation
  */
 
-import { parseTemplateStringE } from "@hoppscotch/data"
 import * as E from "fp-ts/Either"
-import * as O from "fp-ts/Option"
 import * as TE from "fp-ts/TaskEither"
-import { pipe } from "fp-ts/lib/function"
 import { cloneDeep } from "lodash-es"
 
-import { TestDescriptor, TestResponse, TestResult } from "../../types"
-import {
-  getEnv,
-  preventCyclicObjects,
-  setEnv,
-  createExpectation,
-} from "../../utils"
+import { TestResponse, TestResult } from "../../types"
+import { getTestRunnerScriptMethods, preventCyclicObjects } from "../../utils"
 
 const executeScriptInContext = (
   testScript: string,
@@ -23,12 +15,6 @@ const executeScriptInContext = (
   response: TestResponse
 ) => {
   try {
-    let currentEnvs = cloneDeep(envs)
-
-    const testRunStack: TestDescriptor[] = [
-      { descriptor: "root", expectResults: [], children: [] },
-    ]
-
     // Create a function from the script using the Function constructor
     const scriptFunction = new Function(
       "pw",
@@ -40,123 +26,27 @@ const executeScriptInContext = (
       `${testScript}`
     )
 
-    const testFuncHandle = (descriptor: string, testFunc: () => void) => {
-      testRunStack.push({
-        descriptor,
-        expectResults: [],
-        children: [],
-      })
-
-      testFunc()
-
-      const child = testRunStack.pop() as TestDescriptor
-      testRunStack[testRunStack.length - 1].children.push(child)
-    }
-
-    const expectFnHandle = (expectVal: any) =>
-      createExpectation(expectVal, false, testRunStack)
-
-    const envGetHandle = (key: any) => {
-      if (typeof key !== "string") {
-        return TE.left("Expected key to be a string")
-      }
-
-      const result = pipe(
-        getEnv(key, currentEnvs),
-        O.match(
-          () => undefined,
-          ({ value }) => String(value)
-        )
-      )
-
-      return result
-    }
-
-    const envGetResolveHandle = (key: any) => {
-      if (typeof key !== "string") {
-        return TE.left("Expected key to be a string")
-      }
-
-      const result = pipe(
-        getEnv(key, currentEnvs),
-        E.fromOption(() => "INVALID_KEY" as const),
-
-        E.map(({ value }) =>
-          pipe(
-            parseTemplateStringE(value, [...envs.selected, ...envs.global]),
-            // If the recursive resolution failed, return the unresolved value
-            E.getOrElse(() => value)
-          )
-        ),
-        E.map((x) => String(x)),
-
-        E.getOrElseW(() => undefined)
-      )
-
-      return result
-    }
-
-    const envSetHandle = (key: any, value: any) => {
-      if (typeof key !== "string") {
-        return TE.left("Expected key to be a string")
-      }
-
-      if (typeof value !== "string") {
-        return TE.left("Expected value to be a string")
-      }
-
-      currentEnvs = setEnv(key, value, currentEnvs)
-
-      return undefined
-    }
-
-    const envResolveHandle = (value: any) => {
-      if (typeof value !== "string") {
-        return TE.left("Expected value to be a string")
-      }
-
-      const result = pipe(
-        parseTemplateStringE(value, [
-          ...currentEnvs.selected,
-          ...currentEnvs.global,
-        ]),
-        E.getOrElse(() => value)
-      )
-
-      return String(result)
-    }
-
     // Marshal response object
     const responseObjHandle = preventCyclicObjects(response)
     if (E.isLeft(responseObjHandle)) {
       return TE.left(`Response marshalling failed: ${responseObjHandle.left}`)
     }
 
-    const pw = {
-      response: responseObjHandle.right,
-      expect: expectFnHandle,
-      test: testFuncHandle,
-      env: {
-        get: envGetHandle,
-        getResolve: envGetResolveHandle,
-        set: envSetHandle,
-        resolve: envResolveHandle,
-      },
-    }
+    const { pw, testRunStack, updatedEnvs } = getTestRunnerScriptMethods(envs)
 
     // Expose pw and other dependencies to the script
     scriptFunction(
-      pw,
+      { ...pw, response: responseObjHandle.right },
       preventCyclicObjects,
       cloneDeep,
       testRunStack,
-      currentEnvs,
+      updatedEnvs,
       response
     )
 
     return TE.right({
       tests: testRunStack[0],
-      envs: currentEnvs,
+      envs: updatedEnvs,
     })
   } catch (error) {
     return TE.left(`Script execution failed: ${(error as Error).message}`)
